Split API and GUI into multiple files
