+++ dieknow.py
@@ -44,4 +44,6 @@ get_executables_in_folder = lib.get_executables_in_folder
 is_running = lib.is_running
 bsod = lib.bsod
 dialog = lib.dialog
-create_window = lib.create_window
+
+guilib = ctypes.CDLL("./gui.dll")
+create_window = guilib.create_window
