+++ gui.py
@@ -1,0 +1,5 @@
+"""GUI link to Python"""
+
+import dieknow
+
+dieknow.create_window()
